(BSR)[PRO] refactor: move offer table components to component folders
