+++ pro/src/components/CollectiveOffersTable/CollectiveOfferRow/CollectiveActionsCells/BookingLinkCell.tsx
