+++ pro/src/components/CollectiveOffersTable/CollectiveOfferRow/CollectiveActionsCells/CollectiveActionsCells.tsx
@@ -49,14 +49,16 @@ import fullPlusIcon from 'icons/full-plus.svg'
 import fullThreeDotsIcon from 'icons/full-three-dots.svg'
 import strokeCheckIcon from 'icons/stroke-check.svg'
 import strokeThingIcon from 'icons/stroke-thing.svg'
+import styles from 'styles/components/Cells.module.scss'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonLink } from 'ui-kit/Button/ButtonLink'
 import { ButtonVariant } from 'ui-kit/Button/types'
 import { ListIconButton } from 'ui-kit/ListIconButton/ListIconButton'
 
+
 import { BookingLinkCell } from './BookingLinkCell'
-import styles from './Cells.module.scss'
-import { DuplicateOfferDialog } from './DuplicateOfferCell/DuplicateOfferDialog/DuplicateOfferDialog'
+import { DuplicateOfferDialog } from './DuplicateOfferDialog/DuplicateOfferDialog'
+
 
 export interface CollectiveActionsCellsProps {
   offer: CollectiveOfferResponseModel
