+++ pro/src/components/CollectiveOffersTable/CollectiveOfferRow/CollectiveActionsCells/DuplicateOfferDialog/DuplicateOfferDialog.tsx
