+++ pro/src/components/CollectiveOffersTable/CollectiveOfferRow/CollectiveOfferStatusCell/CollectiveOfferStatusCell.tsx
@@ -2,8 +2,7 @@ import cn from 'classnames'
 
 import { CollectiveOfferResponseModel } from 'apiClient/v1'
 import { CollectiveStatusLabel } from 'components/CollectiveStatusLabel/CollectiveStatusLabel'
-
-import styles from '../Cells.module.scss'
+import styles from 'styles/components/Cells.module.scss'
 
 interface CollectiveOfferStatusCellProps {
   offer: CollectiveOfferResponseModel
