+++ pro/src/components/CollectiveOffersTable/CollectiveOfferRow/ExpirationCell/ExpirationCell.module.scss
