+++ pro/src/components/CollectiveOffersTable/CollectiveOfferRow/OfferEventDateCell/OfferEventDateCell.spec.tsx
@@ -3,10 +3,9 @@ import { screen } from '@testing-library/react'
 import { collectiveOfferFactory } from 'commons/utils/factories/collectiveApiFactories'
 import { renderWithProviders } from 'commons/utils/renderWithProviders'
 
-import {
-  OfferEventDateCell,
-  OfferEventDateCellProps,
-} from '../OfferEventDateCell'
+import { OfferEventDateCellProps, OfferEventDateCell } from './OfferEventDateCell'
+
+
 
 const renderOfferNameCell = (props: OfferEventDateCellProps) =>
   renderWithProviders(
