+++ pro/src/components/CollectiveOffersTable/CollectiveOfferRow/OfferEventDateCell/OfferEventDateCell.tsx
@@ -7,8 +7,7 @@ import {
   toDateStrippedOfTimezone,
 } from 'commons/utils/date'
 import { getLocalDepartementDateTimeFromUtc } from 'commons/utils/timezone'
-
-import styles from '../Cells.module.scss'
+import styles from 'styles/components/Cells.module.scss'
 
 export interface OfferEventDateCellProps {
   offer: CollectiveOfferResponseModel
