+++ pro/src/components/OffersTable/Cells/OfferNameCell/__specs__/OfferNameCell.spec.tsx
