+++ pro/src/components/OffersTable/Cells/OfferNameCell/__specs__/utils.offernamecell.spec.tsx
