+++ pro/src/components/OffersTable/Cells/OfferNameCell/utils.ts
