+++ pro/src/components/OffersTable/utils/computeActivationSuccessMessage.ts
