+++ pro/src/components/OffersTable/utils/computeSelectedOffersLabel.ts
