+++ pro/src/pages/Offers/OffersTable/Cells/AddressCell.tsx
@@ -1,9 +1,9 @@
 import classNames from 'classnames'
 
 import { AddressResponseIsLinkedToVenueModel } from 'apiClient/v1'
+import styles from 'styles/components/Cells.module.scss'
 import { computeAddressDisplayName } from 'repository/venuesService'
 
-import styles from './Cells.module.scss'
 
 export function AddressCell({
   address,
