+++ pro/src/pages/Offers/OffersTable/Cells/DeleteDraftCell.tsx
@@ -9,7 +9,6 @@ import {
   OFFER_FORM_NAVIGATION_IN,
   OFFER_FORM_NAVIGATION_MEDIUM,
 } from 'commons/core/FirebaseEvents/constants'
-import { DEFAULT_SEARCH_FILTERS } from 'commons/core/Offers/constants'
 import { useQuerySearchFilters } from 'commons/core/Offers/hooks/useQuerySearchFilters'
 import { useNotification } from 'commons/hooks/useNotification'
 import { ConfirmDialog } from 'components/Dialog/ConfirmDialog/ConfirmDialog'
@@ -20,7 +19,7 @@ import { computeDeletionErrorMessage } from 'pages/Offers/utils/computeDeletionE
 import { computeDeletionSuccessMessage } from 'pages/Offers/utils/computeDeletionSuccessMessage'
 import { ListIconButton } from 'ui-kit/ListIconButton/ListIconButton'
 
-import styles from './Cells.module.scss'
+import styles from 'styles/components/Cells.module.scss'
 import { computeIndividualApiFilters } from 'pages/Offers/utils/computeIndividualApiFilters'
 import { selectCurrentOffererId } from 'commons/store/user/selectors'
 import { useSelector } from 'react-redux'
