+++ pro/src/pages/Offers/OffersTable/Cells/OfferRemainingStockCell.tsx
@@ -4,8 +4,7 @@ import {
   CollectiveOffersStockResponseModel,
   ListOffersStockResponseModel,
 } from 'apiClient/v1'
-
-import styles from './Cells.module.scss'
+import styles from 'styles/components/Cells.module.scss'
 
 const computeRemainingStockValue = (
   stocks: (CollectiveOffersStockResponseModel | ListOffersStockResponseModel)[]
