+++ pro/src/pages/Offers/OffersTable/Cells/OfferStatusCell.tsx
@@ -1,9 +1,9 @@
 import classNames from 'classnames'
 
 import { OfferStatus } from 'apiClient/v1'
+import styles from 'styles/components/Cells.module.scss'
 import { StatusLabel } from 'components/StatusLabel/StatusLabel'
 
-import styles from './Cells.module.scss'
 
 interface OfferStatusCellProps {
   status: OfferStatus
