+++ pro/src/pages/Offers/OffersTable/IndividualOffersTable/IndividualOffersActionsBar/IndividualOffersActionsBar.tsx
@@ -9,17 +9,17 @@ import { serializeApiFilters } from 'commons/core/Offers/utils/serializer'
 import { useActiveFeature } from 'commons/hooks/useActiveFeature'
 import { useNotification } from 'commons/hooks/useNotification'
 import { ActionsBarSticky } from 'components/ActionsBarSticky/ActionsBarSticky'
+import { computeActivationSuccessMessage } from 'components/OffersTable/utils/computeActivationSuccessMessage'
+import { computeSelectedOffersLabel } from 'components/OffersTable/utils/computeSelectedOffersLabel'
 import fullHideIcon from 'icons/full-hide.svg'
 import fullTrashIcon from 'icons/full-trash.svg'
 import fullValidateIcon from 'icons/full-validate.svg'
 import { GET_OFFERS_QUERY_KEY } from 'pages/Offers/OffersRoute'
+import { computeDeletionErrorMessage } from 'pages/Offers/utils/computeDeletionErrorMessage'
 import { computeDeletionSuccessMessage } from 'pages/Offers/utils/computeDeletionSuccessMessage'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonVariant } from 'ui-kit/Button/types'
 
-import { computeActivationSuccessMessage } from '../../../utils/computeActivationSuccessMessage'
-import { computeDeletionErrorMessage } from '../../../utils/computeDeletionErrorMessage'
-import { computeSelectedOffersLabel } from '../../../utils/computeSelectedOffersLabel'
 
 import { DeleteConfirmDialog } from './DeleteConfirmDialog'
 import { IndividualDeactivationConfirmDialog } from './IndividualDeactivationConfirmDialog'
